Move companion app config from sidebar to configuration dashboard
